Non-PreAnnouncement mode for thumbnail generation
Description: These changes are ported from Froyo
Since tiler memory is limited we cannot have 2 instance of decoder 1 instance of encoder
and camera running in parallel. This is a requirement from 1080p camcorder perceptive.
To save tiler memory we are allocating only 2 actual full size buffers but remaining all
buffers as dummy. The dummy buffer will help us to move to idle state. Then we would provide
full allocate buffers only to ducati and then on receiving the Third EmptyThisBufferDone we
would send an EOS. This will flush out the Decoded first I frame buffer back to the component.
This is the frame from which we will make the thumbnail.

Here we have not voilated any OMX protocol but still able to save memory. This is called as
NON-PreAnnoucement mode where we dont give all the actual buffer to the OMX component before
Hand for mapping on the accelerator

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch Set 4: Cherry-picked on current head of tree.  No changes
made.  Taking verifications and approvals from previous patch sets.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: Icd9650f2aa51a2eb919b51962f4c9e61ca763fe5