FM Rx: Added FM framework support
This patch is rework of patch 6235 (cd38325421b73ae5693b1f9991aeeb5b982154c8) on p-froyo for FM Rx
changes. The audio flinger directory is changed in gingerbread. Resubmitting the patch for FM Rx in
gingerbread. All the changes are in OMAP_ENHANCMENT flag

Description: FM Support: Added the FM support in the framework for enabling FM.

Change-Id: I1ff2f6484c673522f83659d9696aba6a2060a798