Gingerbread Camcorder : Fix for color distortion in Recorded Clip
Encoder needs a NULL buffer pointer to be passed to OMX_UseBuffer() fucntion.
Whereas Decoder needs a valid buffer pointer for the same.
This patch adds an overloaded function definition for useBuffer() with extra argument 'size'.
For Encoders, this size is passed to OMX_UseBuffer() along with NULL buffer pointer.
For Decoders, this size is ignored.
With this patch, decision is taken which argument to pass based on if component is Encoder or Decoder.

Signed-off-by : Kalpesh Patil <kalpeshpatil@ti.com>

Patch Set 4: Simple cherry-pick on current tree.  No changes made.
Taking verifications and approvals from patch set 3.
Change-Id: I176881dd632af2f61e9ed6ade02f6d50e1edbc94
Signed-off-by: Jean Johnson <jean-johnson@ti.com>

Patch Set 5: fixed compilation issue with OMAP3
changed #ifdef to be omap4 specific in include/media/IOMX.h