StageFright Test App: Video decoding frame rate support
The changes are for porting the below patch from Froyo to Gingerbread
----------------------------------------------------------
http://review.omapzoom.org/#change,6409

Currently the stagefright test app does the video decoding continuously
without depending on the frame rate.

Have modified the test app to handle the decoding based on frame rate.
It can be passed as input parameter in command line and the clips would
be decoded accordingly.
Eg: "-f 30" would ensure that clip is decoded at 30fps.

Incase this option is not specified it would decode continuously as
earlier.
-----------------------------------------------------------

Change-Id: Icf8c26459b6071323b760f89ecb60a06561e187b
Signed-off-by:<dnk.mohan@ti.com>