OMXCodec changes to handle overlay decided buffer count
The changes are for porting the below patch from Froyo to Gingerbread
----------------------------------------------------------
http://review.omapzoom.org/#change,9211
Description: Currently, the number of buffers required on code output port are
calculated in OMXcodec as well as in TIHardwareRenderer.
And OMXCodec assumes that those many buffers are in deed allocated by overlay.
But in some scenarios it it so possible that, Overlay might get fewer number
of buffers than requested by OMXCodec. In that scenario there will be a
segmentation fault in the system. In order to handle this scenario, OMX codec
is enhanced to reconfigure the codec when it receives the buffers from the
Hardware Renderer (through AwesomePlayer) for the first time.
----------------------------------------------------------

Signed-off-by:<dnk.mohan@ti.com>

Patch Set 3: Cherry-pick on current tree.  No changes made.  Taking
approvals and verifications from patch set 2.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: Ie1642d25eb8f520cb796526882969cb70e751acd