SurfaceFlinger changes to decide when to enable the color key
The changes are for porting the below patch from Froyo to Gingerbread
----------------------------------------------------------
http://review.omapzoom.org/#change,9443
Description: Since color key and per-pixel alpha need to be enabled exclusively
there should be a mechanism in surface flinger to decide when to enable
color key and when to disable.
a new method has been added to the surface flinger to query the graphic plane
pixel format and then enabl/disable color key accordingly.
----------------------------------------------------------

Change-Id: Id7b08c9bcb1187e9eb90fc31f237b3b66c033f4b
Signed-off-by:<dnk.mohan@ti.com>