Awesomeplayer changes for Video Cloning on HDMI after application suspend/resume
The changes are for porting the below patch from Froyo to Gingerbread
----------------------------------------------------------
http://review.omapzoom.org/#change,10447

Description# when Gallery app is suspended by pressing home button, MediaPlayer
gets suspended and underlying Renderer and overlay are destroyed. But the
Awesomeplayer is just reset. Hence the Videomode (normal or Clone) is left
to the last value set before the suspend.
Once the application is resumed, and tries to set the clone mode again,
it would be filtered at the AwesomePlayer level because of the cached value.
This is fixed by resetting the Video mode to normal while resetting the
Awesomeplayer.
----------------------------------------------------------

Signed-off-by:<dnk.mohan@ti.com>

Patch set 2:  Cherry-picked on current tree.  No changes made.
Taking verifications and approvals from patch set 1.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: Ie346c1a81b976ad56cfd7be675a26368556caf35