Stagefright: Handling playback negative seek scenario
The changes are for porting the below patch from Froyo to Gingerbread
----------------------------------------------------------
http://review.omapzoom.org/#change,5711

This patch has change to fix the issue while performing
the stagefright "b 3" tests. They were failing as negative
offset was calculated and used for seek.

Actually in the SF middleware we have condition to check the
parameters. If the seek value is < 0, then android assert is
triggerred which leads to the issue. Actually as per the test app,
the test has to seek and continue decoding till the EOF.
Hence have modified the seek position handling accordingly.
-----------------------------------------------------------

Change-Id: I0fdd8391f1134b91f7c0c03f635de97fb8b7b2cc
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>