UI Cloning support on HDMI and LCD
Description# System server is modified for launching the Ui cloning service.
This service would be listening to HDMI attach/detach events from HDMI
Observer. On attach event it clones the UI to HDTV and on detach event, it
would disable the UI Cloning. And this patch also fixes the hdmi detection
by HDMI observer during bootup. Currently HDMI Obsrever is relying on UEvent from the kernel
for HDMI attach/detach events. But if the HDTV is connected before booting
the platform, the event is lost by the time HDMI observer is initialized
during boot sequence.
This is fixed by modying the HDMIObserver init to read the current display2
status, whether enabled or disabled. Since HDMI HPD event will enable the
display2 device in kernel, this status read is reliable and handles the
hdmi pre-attach event correctly.

Change-Id: I48a63df27796cbb4791fc6bfe98b51dea489d84
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>