OMAP3: create mCodecSpecificData without CodecConfig flag
This fixes a case that encoder returns the first buffer not
including CodecConfig. In case of OMX.TI.Video.encoder using
MPEG4 or H.264, the encoder does not return the CodecConfig
buffer but returns CodecConfig and the first encoded data
in one buffer without OMX_BUFFERFLAG_CODECCONFIG flag set.
Concerning the compatibility of Froyo and GB, this regards only
MPEG4Writer issue in GB.

Change-Id: Ie4d541fd6d3ae540eebbdebed0ef235ec0d557e7
Signed-off-by: keun-o.park <sahara@ti.com>