OMXCodec changes to overload constructor for OMXNode Id
The changes are for porting the below patch from Froyo to Gingerbread
----------------------------------------------------------
http://review.omapzoom.org/#change,10897
Description# OMXCodec create function is overloaded to provide OMXNode Id
handle back. The node Id is used by the Awesomeplayer for configuring the
codec directly.
This patch is only addition of the overloaded method, but the method is
not being used at this point.
----------------------------------------------------------
Patchset2: Fixed build error for OMAP3

Change-Id: I52fb78486048f66e00608e08d31bd20b32f2c178
Signed-off-by:<dnk.mohan@ti.com>