OMAP4:Stagefright player changes to use Codec recommended number of O/P buffers
Description: Currently there is no communication between OMX codec configuration
and the TIHardware renderer wrt the number of Overlay buffers to be used for
on the output port. In this patch, the Awesome player logic is enhanced to
query Codec for the recommended number of buffers for output port and
communicate the same down to TI Hardware Renderer. The codec recommended number
is buffered with the optimal buffer count to be maintained at the DSS driver.
With this patch the hardcoding for the output port buffer number is removed
and also the lower limit for the buffer count is removed in TIHardware Renderer.
this patch has to be merged along with 1.21-p5 ducati migration.

Validation: Tested AV playback from Android Gallery for different resolution cl$

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch set 2: Cherry-pick on current tree.  No changes made.
Taking verifications, approvals and build test from patch set
1.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: I7d858108e8768294ded1b2fc2c2887429f8968a7