UI Cloning Service : Setting correct Z-Order for Overlay on which UI is Cloned
This patch sets Z-Order for Overlay on which UI is Cloned to '0'.

This is needed to avoid following issue:
For Video Playback usecase, when we rotate phone, UI & Video is rotated on LCD.
1)For the portion pertaining to hole in UI-Buffer for Video, Framebuffer ARGB pixel has Alpha= 0
and RGB=000000 (Black)
2)For portion of UI Buffer, where progress-bar appears, Alpha=valid-value and RGB=valid-color.
3)For rest of UI Buffer, where Black-Color appears, Alpha=0xFF and RGB=0x000000 (Black)

But on HDMI-TV, Video (cloned) is not rotated.
This causes mismatch in 'Cloned Framebuffer Layout' and 'Video Layer'.
This causes Black-Color portion of UI (portion-3) to hide Video behind it.

Solution:
For Overlay on which UI is Cloned, Set Z-Order to '0'.
For Overlay of Video-Playback, Set Transparency-Key-Type to 'Gxf-Destination'
& Set Transparency-Key to 0 (Black).
(This should be taken care in TIOverlay.)
With this, Pixels of Cloned UI Buffer with Alpha=0xFF & Black-Color to be hidden behind Video
Overlay and Only Progress-Bar portion of Cloned UI Buffer to come on the top of Video Overlay.

Signed-off-by : Kalpesh Patil <kalpeshpatil@ti.com>

Change-Id: If14145635798a1db3848bac1646f27b168521025