Resolving race condition in AwesomePlayer::reset_l()
In reset_l() mVideoRenderer.clear() is called before mVideoSource->stop(), which means that the overlay output
buffers are being deleted while the OMX component is still in the executing state. Since this is the case there
is no guarantee that the buffer has been returned from the DSP/Ducati before it is deleted. Switching the order
of these two calls guarantees that the buffer will be returned before it is deleted

Patch Set 2:  No changes made.  Taking approvals, verifications and
build test from patch set 1.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>

Change-Id: Id0a056c4b38b4f06a79f9a184a9cbe17bd30d824