Mpeg4 doesnt meet the PPM target
Mpeg4 ppm clips dont meet the validation criteria of 30fps.
The issue was found to be in ctts structure not added to the
stts structure causing wrong time stamp to be presented to
the decoder. This was done only for H264 and not for mpeg4.

DRT: OMAPS00229150
patchset2: fixed review comments.

Change-Id: I71112f48564ad8f778628bfd59bc11c59fcdce6d
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>