AwesomePlayer changes to issue OMXCodec pause
Description# The fillBuffer call in OMXCodec read method is breaking
the mutual exclusion of the out put buffers between Ducati and DSS.
Since we are sending all the buffers not owned by Ducati for filling,
This would include the buffers
(1) that are being rendered by the Display
(2) the ones which are in DSS queue.
(3) the buffers which are in AV sync logic
This results in buffer corruption, hence results in
que failure and tearing effect. This would subsequently result in
DQ buffer failures.

This is fixed by having explicit checks for pause state.
Hence, the read method sends fillbuffer calls only in case
of resume. In normal flow, the fillbuffer is triggered only after
the buffer is signalled from Awesome player.

Validation# Tested AV playback on LCD and HDMI
DRT# OMAPS00233651 - AV playback jitter due to DQ buffer failure

Change-Id: I5e239843d5f8f15d66aa09d8addf9f826f6b6a08
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>