Making MPEG4Writer understand out of order streams
Current mp4 writer doesnt understand out of order frames which is caused
due to insertion of BFrames in codecs

B frames is a bidirection frame and is dependent on the next frame to decode
itself

Current issue is the on turning on the B frames in encoder the timestamps
at the output of the encoder gets organized in a pattern.
e.g of 2 b frames
I/p of encoder I B1 B2 P [0 33 66 99]
O/p of encoder I P B1 B2 [0 99 33 66]

So normal decode timestamp is 99 - 0 , 33 - 66 , 66 - 33
Now we will have negative delta for 33 - 66 which will be encoded.

This is going to cause issue when presented to decoder as we will loose
sign anyways. Standard doesnt allow negative delta encoding in stts structure

So we have to sort the timestamps and then make the stts structure.
Then make a ctts structure which will hold the offset from the decode TS

Signed-off-by: Dandawate Saket <dsaket@ti.com>

Patch set 6: Addressed Dan's comments and cherry-picked on current
tree.  Taking approvals and verifications from patch set 5.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: I1a0eeec83af6cd0e340d29a00bd9f74549eefb90