OMXCodec changes for tearing effect
Description# In the OMXCodec buffer info, a new field has been added to track
the buffers which are with Player. This means the buffer could be in AV sync
logic, or already with DSS for rendering in case of video buffers, or with
Audio HAL incase of Audio buffers. Earlier these buffers are given back
to codec for filling irrespective of whether they are with sink or free.
This is fixed by adding additional state for the buffer to track the
life cycle of the buffer with the sink.

Validation# tested AV playback and AV record.
DRT#OMAPS00234704 - Tearing video frame on display when fastford/rewind
during the video playback

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch Set 2:  Simple cherry-pick on current tree.  No changes made.
Taking approvals, verification and build test from patch set 1.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: I1bcddddfeaf79c624f304b63f0b342625a4d6e33