AwesomePlayer changes to handle the render buffer failure scenario
Description# Currently Awesome player local data structures are updated with
the buffers with renderer after the render call returns. This would leave out
a scenario where, the render call fails to successfully queue the buffer
with the DSS. In this scenario the buffer can't be reclaimed as it is not
registered with the Awesomeplayer.
This is fixed by updating the bufferswithrenderer structure before invoking
the render method.

Validation# AV playback suspend/resume on LCD and HDTV
DRT#OMAPS00234816 - Gallery app stop reponding after pressing END CALL key
while playing a video on HDMI
DRT#OMAPS00234810 - PT: PPM: AV Playback on HDMI. Video is not rendered back
to HDMI after suspend/resume; instead is played distorted in the LCD

Change-Id: Ib6ce15c6f87c81eeb67b4eb6316f8801cdddc816
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>