Stagefright changes to maintain buffer padded W&H along with Video W&H
Description# Stagefright OMXCodec and AwesomePlayer are modified to
store and use the padded width and height for the output buffers.
This is required to handle port reconfiguration scenarios where in,
both video dimensions and buffer dimensions are required. Earlier the OMXCodec
metadata was always holding the padded width and height, hence there was
no way to communicate the video dimensions to Awesomeplayer.

DRT#OMAPS00236082 - Multiple robustness test clips are playing with noise
in the right side of screen due to incorrect port reconfig handling
Note#This is one of the multiple patches for this DR

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch Set 3:  Simple cherry-pick on current head.  Taking approvals,
verfication and build test from patch set 2.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: I441c95dece8b126794f749759d6b6fc651f27094