OMXRenderer changes to propagate display W&H on resizing
Description# OMXRenderer is modified to propagate the display
W&H along with the decoded buffer W&H while resizing the
rendererer. This is required to update the crop window with
Overlay.

Awesomeplayer is modified to query the Video width and height
and provide the same for resize Rendererer call.

DRT#OMAPS00236082 - Multiple robustness test clips are playing with noise
in the right side of screen due to incorrect port reconfig handling
Note#This is one of the multiple patches for this DR
this patch is dependent on
http://review.omapzoom.org/#change,12705

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch Set 3: Simple cherry-pick on current head. Taking approvals,
verification and build test from patch set 2.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: Iceadb97202750b08dd45f8749a5337d5c69f0dc5