webkit: Fixes the browser crash when GPS provider not available
Google websites using GPS were causing Browser crash because Browser
was not handling the case where GPS functionality is not available
on the device. In this case starting of GpslocationService raises the
unhandled exception. Addded the case to handle the exception so that
Browser does not crash when there is problem in GPS service.

Change-Id: I1b7cd854f3a7746c07b5f9a2303af40399b39542
Signed-off-by: Vishal Bhoj <vishal.bhoj@linaro.org>