aapt: update interfaces to libpng 1.5.7 beta 01
Libpng 1.5.7 beta 01 is integrated, where some interfaces are
changed and some structures are not visible to libpng clients
any more. The structure information can be retrieved through
specific interfaces.

The call to png_set_filler() is removed as it can corrupt some png
files. This function was used for libpng 1.2.4, but actually it did
nothing in aapt.

Remove a building warning in ZipFile.h.

Change-Id: I92ca6a2c0d734f9a5187d5a7f6727b8be0e77133