Stagefright: Improve parsing in AH263Assembler
Improves the parsing of the H263 payload header in order to support
playback of more H263 contents. This patch removes check of the

 * RR field which must be ignored by receivers (as specified in
   RFC4629).
 * PLEN field and also modifies the parser to remove any present extra
   picture header (PLEN>0) from the bit stream.
 * PEBIT because its value can be ignored since the extra picture
   header is not forwarded to the decoder anyway.

Note that the extra error resilience provided by an extra picture
header (PLEN>0) is still not utilized because of this patch. The patch
only enables the H263 parser to correctly ignore the extra picture
header.

Change-Id: I1781d9f740cb0c5e15a28bc839d8ce39c6c8fcd0
Signed-off-by: StefanEkenberg <stefan.ekenberg@linaro.org>