surfaceflinger: Silence warnings for missing earlysuspend support.
Earlysuspend is gone from the Android 3.3 kernel but surfaceflinger
still expects it to exist. Apparently this is know about upstream and
fixed, but this hasn't been pushed out into 4.0.x

We don't know what the upstream fix is, so for now we will change
the code to fail silently if the frame-buffer earlysuspend files
are missing.

This fixes LP bug #906210

Change-Id: Id1e30981ac1c94cc59b31515d51e52fbed9e88d9
Signed-off-by: Jon Medhurst <tixy@linaro.org>